@@ -27,29 +27,29 @@
 package jp.fintan.keel.spring.web.token.transaction;
 
 import static org.hamcrest.CoreMatchers.*;
-import static org.junit.Assert.assertThat;
+import static org.hamcrest.MatcherAssert.*;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 /**
  * Test class for TransactionTokenRequestDataValueProcessor
  */
-public class TransactionTokenRequestDataValueProcessorTest {
+class TransactionTokenRequestDataValueProcessorTest {
 
     /**
      * Test method for
-     * {@link TransactionTokenRequestDataValueProcessor#getExtraHiddenFields(javax.servlet.http.HttpServletRequest)}
+     * {@link TransactionTokenRequestDataValueProcessor#getExtraHiddenFields(jakarta.servlet.http.HttpServletRequest)}
      * request has no next token.
      */
     @Test
-    public void testGetExtraHiddenFields01() {
+    void testGetExtraHiddenFields01() {
         // setup arguments
         TransactionTokenRequestDataValueProcessor processor = new TransactionTokenRequestDataValueProcessor();
         HttpServletRequest request = mock(HttpServletRequest.class);
@@ -68,11 +68,11 @@ public class TransactionTokenRequestDataValueProcessorTest {
 
     /**
      * Test method for
-     * {@link TransactionTokenRequestDataValueProcessor#getExtraHiddenFields(javax.servlet.http.HttpServletRequest)}
+     * {@link TransactionTokenRequestDataValueProcessor#getExtraHiddenFields(jakarta.servlet.http.HttpServletRequest)}
      * request has next token.
      */
     @Test
-    public void testGetExtraHiddenFields02() {
+    void testGetExtraHiddenFields02() {
         // setup arguments
         TransactionTokenRequestDataValueProcessor processor = new TransactionTokenRequestDataValueProcessor();
         HttpServletRequest request = mock(HttpServletRequest.class);
