@@ -24,26 +24,15 @@
 package jp.fintan.keel.spring.web.token.transaction;
 
 import static org.hamcrest.CoreMatchers.*;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.hamcrest.MatcherAssert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-public class TransactionTokenTest {
-
-    @Before
-    public void setUp() throws Exception {
-    }
-
-    @After
-    public void tearDown() throws Exception {
-    }
+class TransactionTokenTest {
 
     @Test
-    public void TestTransactionToken01() {
+    void TestTransactionToken01() {
         // setup expected result
         String expectedTokenName = "aaa";
         String expectedTokenKey = "123";
@@ -63,7 +52,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void TestTransactionToken02() {
+    void TestTransactionToken02() {
         // setup expected result
         String expectedTokenName = "aaa";
         String expectedTokenKey = "123";
@@ -83,7 +72,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid01() {
+    void testValid01() {
         // run
         TransactionToken token = new TransactionToken("aaa", "123", "456");
         boolean valid = token.valid();
@@ -93,7 +82,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid02() {
+    void testValid02() {
         // run
         TransactionToken token = new TransactionToken("aaa", "123", "");
         boolean valid = token.valid();
@@ -103,7 +92,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid03() {
+    void testValid03() {
         // run
         TransactionToken token = new TransactionToken("aaa", "123", null);
         boolean valid = token.valid();
@@ -113,7 +102,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid04() {
+    void testValid04() {
         // run
         TransactionToken token = new TransactionToken("aaa", "", "456");
         boolean valid = token.valid();
@@ -123,7 +112,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid05() {
+    void testValid05() {
         // run
         TransactionToken token = new TransactionToken("aaa", null, "456");
         boolean valid = token.valid();
@@ -133,7 +122,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid06() {
+    void testValid06() {
         // run
         TransactionToken token = new TransactionToken("", "123", "456");
         boolean valid = token.valid();
@@ -143,7 +132,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid07() {
+    void testValid07() {
         // run
         TransactionToken token = new TransactionToken(null, "123", "456");
         boolean valid = token.valid();
@@ -153,7 +142,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testValid08() {
+    void testValid08() {
         // run
         TransactionToken token = new TransactionToken(null, null, null);
         boolean valid = token.valid();
@@ -163,7 +152,7 @@ public class TransactionTokenTest {
     }
 
     @Test
-    public void testToString() {
+    void testToString() {
         // run
         TransactionToken token = new TransactionToken("tokenName", "tokenKey", "tokenValue");
         String result = token.getTokenString();
